requirements.txt
pytorch-lightning
onnx
onnxruntime
scipy